codingame_gym/requirements.txt
# Core dependencies
JPype1>=1.4.0
numpy>=1.20.0

# Optional: for running tests
pytest>=7.0.0
pytest-timeout>=2.0.0

# Optional: for hyperparameter tuning
optuna>=3.6.0